+++ fhir/resources/DSTU2/fhirtypesvalidators.py
@@ -12,6 +12,7 @@ from .fhirabstractmodel import FHIRAbstractModel
 __author__ = "Md Nazrul Islam<email2nazrul@gmail.com>"
 
 MODEL_CLASSES = {
+    "FHIRPrimitiveExtension": (None, ".fhirprimitiveextension"),
     "Element": (None, ".element"),
     "Resource": (None, ".resource"),
     "DomainResource": (None, ".domainresource"),
@@ -304,6 +305,11 @@ def fhir_model_validator(
     if model_name != v.resource_type:
         raise ValueError
     return v
+
+
+def fhirprimitiveextension_validator(v: Union[StrBytes, dict, Path, FHIRAbstractModel]):
+
+    return fhir_model_validator("FHIRPrimitiveExtension", v)
 
 
 def element_validator(v: Union[StrBytes, dict, Path, FHIRAbstractModel]):